requirements.txt
# Dependencias para LoL Smart Tracker

Flask>=3.0.3
requests>=2.25
Flask-Caching>=2.0
Flask-SQLAlchemy>=3.0
python-dotenv>=0.20
pandas>=1.3
scikit-learn>=1.0
numpy>=1.20
joblib>=1.1
